feat: add cascading delete for News and its children
- Added middleware in NewsSchema to delete associated comments, replies, and likes when a news item is deleted.

- Enhanced error handling across various parts of the code.

- Refactored code for better simplicity and robustness.

+++ src/models/Comment.js
@@ -47,9 +47,13 @@ const CommentSchema = new mongoose.Schema({
 });
 
 CommentSchema.post('save', async function () {
-    const news = await NewsModel.findById(this.newsId);
-    news.commentCount = this.comment.length;
-    await news.save();
+    try {
+        const news = await NewsModel.findById(this.newsId);
+        news.commentCount = this.comment.length;
+        await news.save();
+    } catch (err) {
+        res.status(500).send({ message: err.message });
+    };
 });
 
 CommentSchema.post('findOneAndUpdate', async function () { await updateCommentCountFromNews(this) });
@@ -57,31 +61,41 @@ CommentSchema.post('findOneAndUpdate', async function () { await updateCommentCo
 CommentSchema.post('updateMany', async function () { await updateCommentCountFromNews(this) });
 
 CommentSchema.pre('updateMany', async function (next) {
-    const { _id: dataCommentId } = this.getQuery();
-    const { $pull: { comment: { _id: commentId } } } = this.getUpdate();
+    try {
+        const { _id: dataCommentId } = this.getQuery();
+        const { $pull: { comment: { _id: commentId } } } = this.getUpdate();
 
-    const comment = await newsService.findCommentById(dataCommentId, commentId);
-    if (comment.comment[0].dataLike) await LikeCommentModel.deleteOne(comment.comment[0].dataLike)
-    if (comment.comment[0].dataReply) {
-        const replies = await ReplyCommentModel.findById(comment.comment[0].dataReply);
-        for (const reply of replies.reply) if (reply.dataLike) await LikeReplyModel.deleteOne(reply.dataLike);
+        const comment = await newsService.findCommentById(dataCommentId, commentId);
+        if (comment.comment[0].dataLike) await LikeCommentModel.deleteOne(comment.comment[0].dataLike)
+        if (comment.comment[0].dataReply) {
+            const replies = await ReplyCommentModel.findById(comment.comment[0].dataReply);
+            if (replies) {
+                for (const reply of replies.reply) if (reply.dataLike) await LikeReplyModel.deleteOne(reply.dataLike);
 
-        await ReplyCommentModel.deleteOne(comment.comment[0].dataReply);
+                await ReplyCommentModel.deleteOne(comment.comment[0].dataReply);
+            }
+        };
+        next();
+    } catch (err) {
+        res.status(500).send({ message: err.message });
     };
-    next();
 });
 
 const updateCommentCountFromNews = async (thisContext) => {
-    const { _id: dataCommentId } = thisContext.getQuery();
+    try {
+        const { _id: dataCommentId } = thisContext.getQuery();
 
-    const getReplyLength = await CommentModel.aggregate([
-        { $match: { _id: dataCommentId } },
-        { $project: { "_id": 0, "newsId": 1, "commentCount": { $size: "$comment" } } }
-    ]);
+        const getReplyLength = await CommentModel.aggregate([
+            { $match: { _id: dataCommentId } },
+            { $project: { "_id": 0, "newsId": 1, "commentCount": { $size: "$comment" } } }
+        ]);
 
-    const news = await NewsModel.findById(getReplyLength[0].newsId);
-    news.commentCount = getReplyLength[0].commentCount;
-    await news.save();
+        const news = await NewsModel.findById(getReplyLength[0].newsId);
+        news.commentCount = getReplyLength[0].commentCount;
+        await news.save();
+    } catch (err) {
+        res.status(500).send({ message: err.message });
+    };
 };
 
 export default CommentSchema;
